% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRel.r
\name{getRel}
\alias{getRel}
\title{Return data from the eurostat and bea APIs}
\usage{
getRel(term = "", lucky = FALSE, beaKey = "")
}
\arguments{
\item{term}{ID of the statistic requested or, if lucky = TRUE, string to search for and return}

\item{lucky}{Boolean operator - if FALSE, must pass relationship ID (see listRel and searchRel)}

\item{beaKey}{Character string representation of user's 36-digit BEA API key}
}
\value{
By default, an object of class 'data.table'
}
\description{
Return data from the eurostat and bea APIs
}

