% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocalMerge.r
\name{loadLocalMerge}
\alias{loadLocalMerge}
\title{Load local merge metadata cache}
\usage{
loadLocalMerge()
}
\value{
Local cache of merge metadata, as data.table
}
\description{
Load local merge metadata cache
}
\keyword{internal}

