% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocalStruc.r
\name{loadLocalStruc}
\alias{loadLocalStruc}
\title{Load local structure metadata cache}
\usage{
loadLocalStruc()
}
\value{
Local cache of structure metadata, as data.table
}
\description{
Load local structure metadata cache
}
\keyword{internal}

