% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRel.R
\name{searchRel}
\alias{searchRel}
\title{Search Metadata Store Relationship Table using text search}
\usage{
searchRel(term, html = FALSE)
}
\arguments{
\item{term}{Search term}

\item{html}{Option to render results in an interactive DT}
}
\description{
Search Metadata Store Relationship Table using text search
}

