% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortlist.R
\name{shortlist}
\alias{shortlist}
\title{Pass list of user specifications (including API key) to return data from BEA API.}
\usage{
shortlist()
}
\value{
By default, an object of class 'data.table' containing a mapping of BEA regional IndustryIDs to NAICS codes
}
\description{
Pass list of user specifications (including API key) to return data from BEA API.
}
\keyword{internal}

