% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLocalStruc.r
\name{updateLocalStruc}
\alias{updateLocalStruc}
\title{Update local structure metadata cache}
\usage{
updateLocalStruc()
}
\value{
Local cache of structure metadata, as data.table
}
\description{
Update local structure metadata cache
}
\keyword{internal}

